\name{k-methods}
\docType{methods}
\alias{k-methods}
\alias{k,ACSPlan-method}
\title{Methods for Function \code{k}}
\description{
Methods for function \code{k} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ACSPlan")}}{
method for extracting critical value \eqn{k} from object of \code{\link{ACSPlan-class}}  (acceptance sampling plan)
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
