\name{n-methods}
\docType{methods}
\alias{n-methods}
\alias{n,ACSPlan-method}
\title{Methods for Function \code{n}}
\description{
 Methods for function \code{n} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ACSPlan")}}{
method for extracting sample size \eqn{n} from object of class \code{\link{ACSPlan-class}} (acceptance sampling plan)
}
}}
\keyword{methods}
