\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ACSPlan-method}
\alias{plot,ANY-method}
\title{Methods for Function \code{plot} in Package \pkg{graphics}}
\description{
Methods for function \code{plot} in package \pkg{graphics}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ACSPlan")}}{
method for plotting \code{\link{OC}} (operating characteristic, i. e. curve showing probability of acceptance of a lot with fraction defective \eqn{p}) for acceptance sampling plan (object of class \code{\link{ACSPlan-class}})
}

\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
